add cloud drop table test
